Enable the inclusion of `frontend.h` from `backend.h`.
