+++ winbolo/src/bolo/frontend.h
@@ -28,8 +28,11 @@
 #define _FRONTEND_H
 
 #include "global.h"
+#include "screen.h"
 #include "screenbullet.h"
+#include "screengunsight.h"
 #include "screenlgm.h"
+#include "screenmines.h"
 #include "screentank.h"
 #include "sound_effects.h"
 #include "types.h"
