+++ winbolo/src/bolo/screengunsight.h
@@ -1,0 +1,23 @@
+/*
+ * Copyright (c) 1998-2008 John Morrison.
+ * Copyright (c) 2024-     Hyrum Wright.
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 2 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ */
+
+/* Defines the gunsight position on the screen */
+/* If turned off mapX is set to NO_GUNSIGHT */
+typedef struct {
+  int mapX;
+  BYTE mapY;
+  BYTE pixelX;
+  BYTE pixelY;
+} screenGunsight;
